Testsuite in der Linux-Sandbox lauffaehig: 343 Tests gruen
Die Gegenprobe aus Abschnitt 17/18 galt als nicht durchfuehrbar, weil pytest
in der Sandbox nicht installierbar ist. Fast alle Testabhaengigkeiten liegen
aber als reines Python im Windows-venv; nur exceptiongroup fehlte (Python 3.10
kennt BaseExceptionGroup nicht). werkzeuge/pytest_linux.py sammelt beides nach
/tmp und startet pytest darueber.

Ergebnis: 343 Tests gruen. Die Aenderungen aus Abschnitt 17 brechen nichts.

Nebenbefunde: die Testzahl in CLAUDE.md war mit 337 veraltet; ein FutureWarning
in tests/test_metrics_and_splits.py (ffill auf object-dtype) ist beseitigt.

Der Lauf laeuft unter Python 3.10 mit Linux-pandas und ersetzt den
Windows-Lauf nicht -- in CLAUDE.md und Abschnitt 20 so gekennzeichnet.

+++ tests/test_metrics_and_splits.py
@@ -34,7 +34,9 @@ def make_trade(pnl: float, *, index: int = 0, bars: int = 10) -> Trade:
 def make_result(pnls: list[float]) -> BacktestResult:
     index = pd.date_range("2025-01-02", periods=max(len(pnls), 2), freq="1D", tz="UTC")
     equity = pd.Series(
-        pd.Series(pnls).cumsum().reindex(range(len(index))).ffill().fillna(0.0).values,
+        # dtype ausdruecklich: bei leerer Liste waere die Reihe sonst object-dtype,
+        # und ffill auf object-dtype ist in pandas abgekuendigt (FutureWarning).
+        pd.Series(pnls, dtype="float64").cumsum().reindex(range(len(index))).ffill().fillna(0.0).values,
         index=index,
     )
     return BacktestResult(
